# Optimize SummaryValue for speed