This change removes unnecessary Snapshot message and replaces wrapped int and double
messages by regular primitive data types. Also removed count and sum from Snapshot
because there are no known use-cases for these.

Benchmark in Go demonstrates the following improvement of encoding and decoding
compared to the baseline state:

```
BenchmarkEncode/Baseline/Metric/Summary-8      	      84	  64514181 ns/op
BenchmarkEncode/Proposed/Metric/Summary-8      	     100	  52800440 ns/op

BenchmarkDecode/Baseline/Metric/Summary-8      	      39	 146303714 ns/op	95496042 B/op	 2324000 allocs/op
BenchmarkDecode/Proposed/Metric/Summary-8      	      52	 120249385 ns/op	79496032 B/op	 2024000 allocs/op
```

Encoding and decoding is about 18% faster, in-memory representation is about 16% smaller.

Benchmark source code is available at:
https://github.com/tigrannajaryan/exp-otelproto/blob/70f5b075fda85f115102798eae3a7a9f7fedb5e5/encodings/encoding_test.go