@tigrannajaryan 2019-11-08
> Please mention that this is compatible with OM

OpenMetrics proto exists only as a PR. It is not merged into their master branch. Is it useful to refer to a PR that may not end up being merged? By the time it is merged it may be changed again and our claim to be compatible with OM may be wrong.